% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{matrixplot}
\alias{matrixplot}
\title{Plot a matrix as an heatmap}
\usage{
matrixplot(m1, m2 = NULL)
}
\arguments{
\item{m1}{the matrix to be plotted}

\item{m2}{(optional) second matrix to be plotted next to \code{m1}. Can be used to compare estimated vs true or to plot separately endogenous and exogenous covariates effects}
}
\description{
Plot a matrix as an heatmap
}
