% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{qDensity}
\alias{qDensity}
\title{Plot the optimal densities for \beta and \theta}
\usage{
qDensity(mod, param, index)
}
\arguments{
\item{mod}{an output of \code{VBmlr}}

\item{param}{the parameter to be plotted: either "beta" or "theta"}

\item{index}{the index of the parameter, i.e., the entry in the corresponding matrix \beta or \theta}
}
\value{
95% HPD interval and the density plot
}
\description{
Plot the optimal densities for \beta and \theta
}
