% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{qExactPredictive}
\alias{qExactPredictive}
\title{Exact variational predictive distribution (see the paper for more details). Only for \code{SV=FALSE}.}
\usage{
qExactPredictive(mod, qOm, z, R = 5000)
}
\arguments{
\item{mod}{an output of \code{VBmlr} with \code{SV=FALSE}}

\item{qOm}{an output of \code{qOmega}}

\item{z}{the vector of new observed covariates for prediction}

\item{R}{the number of samples from the variational predictive posterior}
}
\value{
Samples from the variational predictive posterior \code{d x R}
}
\description{
Exact variational predictive distribution (see the paper for more details). Only for \code{SV=FALSE}.
}
