% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{qOmega}
\alias{qOmega}
\title{Density approximation for Omega (see the paper for more details). Only for \code{SV=FALSE}.}
\usage{
qOmega(mod)
}
\arguments{
\item{mod}{an output of \code{VBmlr} with \code{SV=FALSE}}
}
\value{
\code{df}: degrees of freedom of the Wishart approximation

\code{V}: scale matrix of the Wishart approximation
}
\description{
Density approximation for Omega (see the paper for more details). Only for \code{SV=FALSE}.
}
