% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ran_mlr}
\alias{ran_mlr}
\title{Generate samples from the model}
\usage{
ran_mlr(n, params, nburn = 1000L, seed_set = 0L)
}
\arguments{
\item{n}{number of samples}

\item{params}{list of parameters: 1) \code{Omega}: the precision matrix (\code{d x d}), 2) \code{mu}: the intercept vector (\code{d x 1}), 3) \code{Phi}: the autoregressive matrix (\code{d x d}), 4) \code{Gamma}: the regression coefficients matrix (\code{d x p})}

\item{nburn}{number of observations to discard before sampling}

\item{seed_set}{set the seed for replicability (if 0, no set seed)}
}
\value{
\code{Y} the matrix of observations (\code{d x n})

\code{X} the matrix of exogenous covariates (independent normally distributed) (\code{p x n})
}
\description{
Generate samples from the model
}
