\name{vbmlr-package}
\alias{vbmlr-package}
\alias{vbmlr}
\docType{package}
\title{
\packageTitle{vbmlr}
}
\description{
\packageDescription{vbmlr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vbmlr}
\packageIndices{vbmlr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{vbmlr}

Maintainer: \packageMaintainer{vbmlr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
