% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{VBmlr}
\alias{VBmlr}
\title{Variational Bayes inference for multivariate linear regressions with autoregressive effects and stochastic volatility (Bernardi, Bianchi, and Bianco, 2024: Variational Inference for Large Bayesian Vector Autoregressions)}
\usage{
VBmlr(
  Y,
  X = NULL,
  AR = TRUE,
  hyper = list(a_nu = 0.1, b_nu = 0.1, tau = 1, ups = 1),
  prior = "normal",
  SV = FALSE,
  maxIter = 500,
  Tol_ELBO = 0.01,
  Tol_Par = 0.01,
  Trace = 0
)
}
\arguments{
\item{Y}{is a \code{d} times \code{T} matrix of observations}

\item{X}{is a \code{p} times \code{T} matrix of exogenous covariates. If not specified (i.e, \code{X=NULL}) only an intercept is included}

\item{AR}{if \code{TRUE} an autoregressive effect of order 1 is included (VAR(1))}

\item{hyper}{is a list of algorithm hyperparameters depending on \code{prior}}

\item{prior}{can be "normal" (gaussian prior), "lasso" (adaptive Bayesian lasso), "ng" (normal-gamma), "hs" (horseshoe)}

\item{SV}{if \code{TRUE} stochastic volatility is added to the model}

\item{maxIter}{maximum number of iterations in the CAVI algorithm}

\item{Tol_ELBO}{tolerance for relative variation in ELBO for convergence}

\item{Tol_Par}{tolerance for relative variation in variational parameters for convergence}

\item{Trace}{if 1 display the progress of the algorithm}
}
\value{
\code{Omega_hat}: Estimated precision matrix (posterior mean of the approximation), \code{d x d} if the \code{SV=FALSE} and \code{d x d x n} if the \code{SV=TRUE}

\code{a_q_nu}: Shape parameter of the posterior gamma approximation for \nu (only for \code{SV=FALSE})

\code{b_q_nu}: Rate parameter of the posterior gamma approximation for \nu (only for \code{SV=FALSE})

\code{mu_q_nu}: Posterior mean of the gamma approximation for \nu

\code{Mu_q_beta}: Posterior mean of the normal approximation for \beta (\code{d x d})

\code{Sigma_q_beta}: Posterior variances of the normal approximation for \beta (\code{d x d x d})

\code{Mu_q_theta}: Posterior mean of the normal approximation for \theta (\code{d x (d+p+1)}). It includes autoregressive matrix if \code{AR=TRUE}, intercept and regression coefficient matrix. See the paper for details.

\code{Sigma_q_theta}: Posterior variances of the normal approximation for \theta (\code{(d+p+1) x (d+p+1) x d}). It includes autoregressive matrix if \code{AR=TRUE}, intercept and regression coefficient matrix. See the paper for details.

\code{lowerBoundIter}: collection of ELBO until convergence

\code{lowerBound}: value of ELBO at convergence

\code{convergence}: check: if 1 successfully converged
}
\description{
Variational Bayes inference for multivariate linear regressions with autoregressive effects and stochastic volatility (Bernardi, Bianchi, and Bianco, 2024: Variational Inference for Large Bayesian Vector Autoregressions)
}
